requirements.txt
ultralytics
opencv-python
numpy
moviepy==1.0.3
imageio-ffmpeg
requests
pillow
# Local Vision Models
transformers
torch
torchvision
sentence-transformers
# Optional: for faster inference
accelerate
# Audio processing
librosa
soundfile
# Flask API Server
flask
flask-cors
werkzeug
python-dotenv
